+++ backend/src/main/kotlin/com/woowahan/campus/zzimkkong/feature/campus/CreateCampus.kt
@@ -1,4 +1,4 @@
-package com.woowahan.campus.zzimkkong.feature
+package com.woowahan.campus.zzimkkong.feature.campus
 
 import com.woowahan.campus.zzimkkong.domain.Campus
 import com.woowahan.campus.zzimkkong.domain.CampusRepository
