+++ backend/src/main/kotlin/com/woowahan/campus/zzimkkong/feature/space/UpdateSpace.kt
@@ -1,45 +1,62 @@
-package com.woowahan.campus.zzimkkong.feature
+package com.woowahan.campus.zzimkkong.feature.space
 
 import com.woowahan.campus.zzimkkong.domain.CampusRepository
 import com.woowahan.campus.zzimkkong.domain.Setting
 import com.woowahan.campus.zzimkkong.domain.SettingRepository
 import com.woowahan.campus.zzimkkong.domain.Space
 import com.woowahan.campus.zzimkkong.domain.SpaceRepository
-import openapi.api.CreateSpaceApi
-import openapi.model.SpacePost
-import openapi.model.SpacePostSettingsInnerEnabledDayOfWeek
+import openapi.api.UpdateSpaceApi
+import openapi.model.SpacePut
+import openapi.model.SpacePutSettingsInner
+import openapi.model.SpacePutSettingsInnerEnabledDayOfWeek
 import org.springframework.http.ResponseEntity
 import org.springframework.transaction.annotation.Transactional
 import org.springframework.web.bind.annotation.RestController
-import java.net.URI
 import java.time.DayOfWeek
 import java.time.LocalTime
 
 @RestController
-class CreateSpace(
-    val campusRepository: CampusRepository,
+class UpdateSpace(
     val spaceRepository: SpaceRepository,
     val settingRepository: SettingRepository,
-) : CreateSpaceApi {
+    val campusRepository: CampusRepository,
+) : UpdateSpaceApi {
 
     @Transactional
-    override fun createSpace(mapId: Int, spacePost: SpacePost): ResponseEntity<Unit> {
+    override fun updateSpace(mapId: Int, spaceId: Int, spacePut: SpacePut): ResponseEntity<Unit> {
+        updateThumbnail(mapId, spacePut.thumbnail)
+        val findSpace = updateSpace(spaceId, spacePut)
+        updateSettings(findSpace, spacePut.settings)
+
+        return ResponseEntity.ok().build()
+    }
+
+    private fun updateThumbnail(mapId: Int, thumbnail: String) {
         val campus = campusRepository.findById(mapId.toLong())
             .orElseThrow { IllegalArgumentException() }
+        campus.updateThumbnail(thumbnail)
+        campusRepository.save(campus)
+    }
 
-        val space = spaceRepository.save(
-            Space(
-                campusId = campus.id,
-                name = spacePost.name,
-                color = spacePost.color,
-                area = spacePost.area,
-                reservationEnabled = spacePost.reservationEnable,
-            )
+    private fun updateSpace(spaceId: Int, spacePut: SpacePut): Space {
+        val findSpace = spaceRepository.findById(spaceId.toLong())
+            .orElseThrow { IllegalArgumentException() }
+        findSpace.update(
+            name = spacePut.name,
+            color = spacePut.color,
+            area = spacePut.area,
+            reservationEnabled = spacePut.reservationEnable,
         )
+        return spaceRepository.save(findSpace)
+    }
+
+    private fun updateSettings(findSpace: Space, settings: List<SpacePutSettingsInner>) {
+        settingRepository.deleteAllBySpaceId(findSpace.id)
+
         settingRepository.saveAll(
-            spacePost.settings.map {
+            settings.map {
                 Setting(
-                    spaceId = space.id,
+                    spaceId = findSpace.id,
                     startTime = LocalTime.parse(it.settingStartTime),
                     endTime = LocalTime.parse(it.settingEndTime),
                     maximumMinute = it.reservationMaximumTimeUnit,
@@ -47,14 +64,9 @@ class CreateSpace(
                 )
             }.toList()
         )
-
-        campus.updateThumbnail(spacePost.thumbnail)
-        campusRepository.save(campus)
-
-        return ResponseEntity.created(URI.create("/api/maps/${campus.id}/spaces/${space.id}")).build()
     }
 
-    private fun parseToEnableDays(it: SpacePostSettingsInnerEnabledDayOfWeek): String {
+    private fun parseToEnableDays(it: SpacePutSettingsInnerEnabledDayOfWeek): String {
         return DayOfWeek.values()
             .filter { day ->
                 when (day) {
