+++ backend/src/test/kotlin/com/woowahan/campus/zzimkkong/feature/campus/CreateCampusTest.kt
@@ -1,4 +1,4 @@
-package com.woowahan.campus.zzimkkong.feature
+package com.woowahan.campus.zzimkkong.feature.campus
 
 import com.woowahan.campus.zzimkkong.fixture.CampusFixture
 import io.kotest.core.spec.style.BehaviorSpec
